pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "moviepy==1.0.3",
    "pdf2docx>=0.5.8",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "reportlab>=4.4.2",
    "telegram>=0.0.1",
]
